(PUP-7848) Modify var name regexp to avoid infinite recursion
The VAR_NAME regular expression caused an infinite left recursion due to
an ambiguity in this regular expression. This manifested itself when the
trailing part of a FQN did not match the trailing part of the regexp.
For example `$foo::bar::WRONG`

This commit changes the regexp so that left recursion is not needed.