(PUP-7848) Fix problem with VAR pattern and longer segments
This is yet another fix of the same problem as earlier believed to be
fixed. When input such as $aaaaa::bbbbb::ccccc::UPPER was encountered
the regexp matcher for variable/name ended up taking an enormous amount
of time. The initial regexp triggered this problem on very short
sequences. The updated worked for short sequences, but not for longer
(also took a very long time to complete).

The regexp is now updated with a more complex negative lookahead of '_'
as the problem is to prevent that a segment other than the last starts
with '_'.

Tests are updated to have longer sequences, and also includes the
original reported name.