Do not process Node definitions
There seems to be a bug allowing default Node definitions to be leaked
across catalogs. If a node A has a:

  node default {
    resources here
  }

in the catalog and a thread compiles for it, a node B coming behind will
pick up this definition which is bad in a "shared master" environment.
Surprisingly enough this only happens during the immediate following
compilation to A's. If B requests yet another catalog it will be clean.

In our environment (where there's an ENC) we don't want people to even
use Node declarations so as a workaround for this bug we're now
completely ignoring these at the time of walking the AST.